Fix project service/project participation service tests
