+++ api/src/paths/project/{projectId}/participants/{projectParticipationId}/update.test.ts
@@ -4,9 +4,8 @@ import sinon from 'sinon';
 import sinonChai from 'sinon-chai';
 import * as db from '../../../../../database/db';
 import { HTTPError } from '../../../../../errors/http-error';
-import { ProjectService } from '../../../../../services/project-service';
+import { ProjectParticipationService } from '../../../../../services/project-participation-service';
 import { getMockDBConnection, getRequestHandlerMocks } from '../../../../../__mocks__/db';
-import * as doAllProjectsHaveAProjectLead from '../../../../user/{userId}/delete';
 import * as update_project_participant from './update';
 chai.use(sinonChai);
 
@@ -22,9 +21,9 @@ describe('update a project participant.', () => {
     mockReq.params = { projectId: '1', projectParticipationId: '2' };
     mockReq.body = { roleId: '1' };
 
-    sinon.stub(ProjectService.prototype, 'deleteProjectParticipationRecord').resolves();
-    sinon.stub(ProjectService.prototype, 'getProjectParticipants').resolves([{ id: 1 }]);
-    sinon.stub(doAllProjectsHaveAProjectLead, 'doAllProjectsHaveAProjectLead').returns(true);
+    sinon.stub(ProjectParticipationService.prototype, 'deleteProjectParticipationRecord').resolves();
+    sinon.stub(ProjectParticipationService.prototype, 'getProjectParticipants').resolves([{ id: 1 }]);
+    sinon.stub(ProjectParticipationService.prototype, 'doAllProjectsHaveAProjectLead').returns(true);
 
     sinon.stub(db, 'getDBConnection').returns({
       ...dbConnectionObj,
@@ -51,10 +50,12 @@ describe('update a project participant.', () => {
     mockReq.params = { projectId: '1', projectParticipationId: '2' };
     mockReq.body = { roleId: '1' };
 
-    sinon.stub(ProjectService.prototype, 'deleteProjectParticipationRecord').resolves({ system_user_id: 1 });
-    sinon.stub(ProjectService.prototype, 'addProjectParticipant').resolves();
-    const getProjectParticipant = sinon.stub(ProjectService.prototype, 'getProjectParticipants');
-    const doAllProjectsHaveLead = sinon.stub(doAllProjectsHaveAProjectLead, 'doAllProjectsHaveAProjectLead');
+    sinon
+      .stub(ProjectParticipationService.prototype, 'deleteProjectParticipationRecord')
+      .resolves({ system_user_id: 1 });
+    sinon.stub(ProjectParticipationService.prototype, 'addProjectParticipant').resolves();
+    const getProjectParticipant = sinon.stub(ProjectParticipationService.prototype, 'getProjectParticipants');
+    const doAllProjectsHaveLead = sinon.stub(ProjectParticipationService.prototype, 'doAllProjectsHaveAProjectLead');
 
     getProjectParticipant.onCall(0).resolves([{ id: 1 }]);
     doAllProjectsHaveLead.onCall(0).returns(true);
@@ -88,10 +89,12 @@ describe('update a project participant.', () => {
     mockReq.params = { projectId: '1', projectParticipationId: '2' };
     mockReq.body = { roleId: '1' };
 
-    sinon.stub(ProjectService.prototype, 'deleteProjectParticipationRecord').resolves({ system_user_id: 1 });
-    sinon.stub(ProjectService.prototype, 'addProjectParticipant').resolves();
-    const getProjectParticipant = sinon.stub(ProjectService.prototype, 'getProjectParticipants');
-    const doAllProjectsHaveLead = sinon.stub(doAllProjectsHaveAProjectLead, 'doAllProjectsHaveAProjectLead');
+    sinon
+      .stub(ProjectParticipationService.prototype, 'deleteProjectParticipationRecord')
+      .resolves({ system_user_id: 1 });
+    sinon.stub(ProjectParticipationService.prototype, 'addProjectParticipant').resolves();
+    const getProjectParticipant = sinon.stub(ProjectParticipationService.prototype, 'getProjectParticipants');
+    const doAllProjectsHaveLead = sinon.stub(ProjectParticipationService.prototype, 'doAllProjectsHaveAProjectLead');
 
     getProjectParticipant.onCall(0).resolves([{ id: 1 }]);
     doAllProjectsHaveLead.onCall(0).returns(true);
