+++ api/src/paths/user/{userId}/delete.test.ts
@@ -77,7 +77,7 @@ describe('removeSystemUser', () => {
 
     sinon.stub(db, 'getDBConnection').returns(dbConnectionObj);
 
-    sinon.stub(delete_endpoint, 'checkIfUserIsOnlyProjectLeadOnAnyProject').resolves();
+    sinon.stub(ProjectParticipationService.prototype, 'checkIfUserIsOnlyProjectLeadOnAnyProject').resolves();
 
     sinon.stub(UserService.prototype, 'getUserById').resolves({
       system_user_id: 1,
@@ -114,7 +114,7 @@ describe('removeSystemUser', () => {
 
     sinon.stub(db, 'getDBConnection').returns(dbConnectionObj);
 
-    sinon.stub(delete_endpoint, 'checkIfUserIsOnlyProjectLeadOnAnyProject').resolves();
+    sinon.stub(ProjectParticipationService.prototype, 'checkIfUserIsOnlyProjectLeadOnAnyProject').resolves();
 
     sinon.stub(UserService.prototype, 'getUserById').resolves({
       system_user_id: 1,
@@ -152,7 +152,7 @@ describe('removeSystemUser', () => {
 
     sinon.stub(db, 'getDBConnection').returns(dbConnectionObj);
 
-    sinon.stub(delete_endpoint, 'checkIfUserIsOnlyProjectLeadOnAnyProject').resolves();
+    sinon.stub(ProjectParticipationService.prototype, 'checkIfUserIsOnlyProjectLeadOnAnyProject').resolves();
 
     sinon.stub(UserService.prototype, 'getUserById').resolves({
       system_user_id: 1,
@@ -192,7 +192,7 @@ describe('removeSystemUser', () => {
 
     sinon.stub(db, 'getDBConnection').returns(dbConnectionObj);
 
-    sinon.stub(delete_endpoint, 'checkIfUserIsOnlyProjectLeadOnAnyProject').resolves();
+    sinon.stub(ProjectParticipationService.prototype, 'checkIfUserIsOnlyProjectLeadOnAnyProject').resolves();
 
     sinon.stub(UserService.prototype, 'getUserById').resolves({
       system_user_id: 1,
@@ -233,7 +233,7 @@ describe('removeSystemUser', () => {
 
     sinon.stub(db, 'getDBConnection').returns(dbConnectionObj);
 
-    sinon.stub(delete_endpoint, 'checkIfUserIsOnlyProjectLeadOnAnyProject').resolves();
+    sinon.stub(ProjectParticipationService.prototype, 'checkIfUserIsOnlyProjectLeadOnAnyProject').resolves();
 
     sinon.stub(UserService.prototype, 'getUserById').resolves({
       system_user_id: 1,
@@ -257,361 +257,5 @@ describe('removeSystemUser', () => {
     await requestHandler(mockReq, mockRes, mockNext);
 
     expect(mockRes.statusValue).to.equal(200);
-  });
-});
-
-describe('doAllProjectsHaveAProjectLeadIfUserIsRemoved', () => {
-  describe('user has Coordinator role', () => {
-    describe('user is on 1 project', () => {
-      it('should return false if the user is not the only Coordinator role', () => {
-        const userId = 10;
-
-        const rows = [
-          {
-            project_participation_id: 1,
-            project_id: 1,
-            system_user_id: userId,
-            project_role_id: 1,
-            project_role_name: 'Coordinator'
-          },
-          {
-            project_participation_id: 2,
-            project_id: 1,
-            system_user_id: 20,
-            project_role_id: 1,
-            project_role_name: 'Coordinator'
-          }
-        ];
-
-        const result = delete_endpoint.doAllProjectsHaveAProjectLeadIfUserIsRemoved(rows, userId);
-
-        expect(result).to.equal(true);
-      });
-
-      it('should return true if the user is the only Coordinator role', () => {
-        const userId = 10;
-
-        const rows = [
-          {
-            project_participation_id: 1,
-            project_id: 1,
-            system_user_id: userId,
-            project_role_id: 1,
-            project_role_name: 'Coordinator' // Only Coordinator on project 1
-          },
-          {
-            project_participation_id: 2,
-            project_id: 1,
-            system_user_id: 20,
-            project_role_id: 2,
-            project_role_name: 'Collaborator'
-          }
-        ];
-
-        const result = delete_endpoint.doAllProjectsHaveAProjectLeadIfUserIsRemoved(rows, userId);
-
-        expect(result).to.equal(false);
-      });
-    });
-
-    describe('user is on multiple projects', () => {
-      it('should return true if the user is not the only Coordinator on all projects', () => {
-        const userId = 10;
-
-        const rows = [
-          {
-            project_participation_id: 1,
-            project_id: 1,
-            system_user_id: userId,
-            project_role_id: 1,
-            project_role_name: 'Coordinator'
-          },
-          {
-            project_participation_id: 2,
-            project_id: 1,
-            system_user_id: 2,
-            project_role_id: 1,
-            project_role_name: 'Coordinator'
-          },
-          {
-            project_participation_id: 1,
-            project_id: 2,
-            system_user_id: userId,
-            project_role_id: 1,
-            project_role_name: 'Coordinator'
-          },
-          {
-            project_participation_id: 2,
-            project_id: 2,
-            system_user_id: 2,
-            project_role_id: 1,
-            project_role_name: 'Coordinator'
-          }
-        ];
-
-        const result = delete_endpoint.doAllProjectsHaveAProjectLeadIfUserIsRemoved(rows, userId);
-
-        expect(result).to.equal(true);
-      });
-
-      it('should return false if the user the only Coordinator on any project', () => {
-        const userId = 10;
-
-        // User is on 1 project, and is not the only Coordinator
-        const rows = [
-          {
-            project_participation_id: 1,
-            project_id: 1,
-            system_user_id: userId,
-            project_role_id: 1,
-            project_role_name: 'Collaborator'
-          },
-          {
-            project_participation_id: 2,
-            project_id: 1,
-            system_user_id: 2,
-            project_role_id: 1,
-            project_role_name: 'Coordinator'
-          },
-          {
-            project_participation_id: 1,
-            project_id: 2,
-            system_user_id: userId,
-            project_role_id: 1,
-            project_role_name: 'Coordinator' // Only Coordinator on project 2
-          },
-          {
-            project_participation_id: 2,
-            project_id: 2,
-            system_user_id: 2,
-            project_role_id: 1,
-            project_role_name: 'Collaborator'
-          }
-        ];
-
-        const result = delete_endpoint.doAllProjectsHaveAProjectLeadIfUserIsRemoved(rows, userId);
-
-        expect(result).to.equal(false);
-      });
-    });
-  });
-
-  describe('user does not have Coordinator role', () => {
-    describe('user is on 1 project', () => {
-      it('should return true', () => {
-        const userId = 10;
-
-        const rows = [
-          {
-            project_participation_id: 1,
-            project_id: 1,
-            system_user_id: userId,
-            project_role_id: 1,
-            project_role_name: 'Collaborator'
-          },
-          {
-            project_participation_id: 2,
-            project_id: 1,
-            system_user_id: 20,
-            project_role_id: 1,
-            project_role_name: 'Coordinator'
-          }
-        ];
-
-        const result = delete_endpoint.doAllProjectsHaveAProjectLeadIfUserIsRemoved(rows, userId);
-
-        expect(result).to.equal(true);
-      });
-    });
-
-    describe('user is on multiple projects', () => {
-      it('should return true', () => {
-        const userId = 10;
-
-        const rows = [
-          {
-            project_participation_id: 1,
-            project_id: 1,
-            system_user_id: userId,
-            project_role_id: 1,
-            project_role_name: 'Collaborator'
-          },
-          {
-            project_participation_id: 2,
-            project_id: 1,
-            system_user_id: 2,
-            project_role_id: 1,
-            project_role_name: 'Coordinator'
-          },
-          {
-            project_participation_id: 1,
-            project_id: 2,
-            system_user_id: userId,
-            project_role_id: 1,
-            project_role_name: 'Observer'
-          },
-          {
-            project_participation_id: 2,
-            project_id: 2,
-            system_user_id: 2,
-            project_role_id: 1,
-            project_role_name: 'Coordinator'
-          }
-        ];
-
-        const result = delete_endpoint.doAllProjectsHaveAProjectLeadIfUserIsRemoved(rows, userId);
-
-        expect(result).to.equal(true);
-      });
-    });
-  });
-
-  describe('user is on no projects', () => {
-    it('should return false', () => {
-      const userId = 10;
-
-      const rows = [
-        {
-          project_participation_id: 1,
-          project_id: 1,
-          system_user_id: 20,
-          project_role_id: 1,
-          project_role_name: 'Collaborator'
-        },
-        {
-          project_participation_id: 2,
-          project_id: 1,
-          system_user_id: 30,
-          project_role_id: 1,
-          project_role_name: 'Coordinator'
-        }
-      ];
-
-      const result = delete_endpoint.doAllProjectsHaveAProjectLeadIfUserIsRemoved(rows, userId);
-
-      expect(result).to.equal(true);
-    });
-  });
-});
-
-describe('doAllProjectsHaveAProjectLead', () => {
-  it('should return false if no user has Coordinator role', () => {
-    const rows = [
-      {
-        project_participation_id: 1,
-        project_id: 1,
-        system_user_id: 10,
-        project_role_id: 2,
-        project_role_name: 'Collaborator'
-      },
-      {
-        project_participation_id: 2,
-        project_id: 1,
-        system_user_id: 20,
-        project_role_id: 2,
-        project_role_name: 'Collaborator'
-      }
-    ];
-
-    const result = delete_endpoint.doAllProjectsHaveAProjectLead(rows);
-
-    expect(result).to.equal(false);
-  });
-
-  it('should return true if one Coordinator role exists per project', () => {
-    const rows = [
-      {
-        project_participation_id: 1,
-        project_id: 1,
-        system_user_id: 12,
-        project_role_id: 1,
-        project_role_name: 'Coordinator' // Only Coordinator on project 1
-      },
-      {
-        project_participation_id: 2,
-        project_id: 1,
-        system_user_id: 20,
-        project_role_id: 2,
-        project_role_name: 'Collaborator'
-      }
-    ];
-
-    const result = delete_endpoint.doAllProjectsHaveAProjectLead(rows);
-
-    expect(result).to.equal(true);
-  });
-
-  it('should return true if one Coordinator exists on all projects', () => {
-    const rows = [
-      {
-        project_participation_id: 1,
-        project_id: 1,
-        system_user_id: 10,
-        project_role_id: 1,
-        project_role_name: 'Coordinator'
-      },
-      {
-        project_participation_id: 2,
-        project_id: 1,
-        system_user_id: 2,
-        project_role_id: 2,
-        project_role_name: 'Collaborator'
-      },
-      {
-        project_participation_id: 1,
-        project_id: 2,
-        system_user_id: 10,
-        project_role_id: 1,
-        project_role_name: 'Coordinator'
-      },
-      {
-        project_participation_id: 2,
-        project_id: 2,
-        system_user_id: 2,
-        project_role_id: 2,
-        project_role_name: 'Collaborator'
-      }
-    ];
-
-    const result = delete_endpoint.doAllProjectsHaveAProjectLead(rows);
-
-    expect(result).to.equal(true);
-  });
-
-  it('should return false if no Coordinator exists on any one project', () => {
-    const rows = [
-      {
-        project_participation_id: 1,
-        project_id: 1,
-        system_user_id: 10,
-        project_role_id: 1,
-        project_role_name: 'Coordinator'
-      },
-      {
-        project_participation_id: 2,
-        project_id: 1,
-        system_user_id: 20,
-        project_role_id: 2,
-        project_role_name: 'Collaborator'
-      },
-      {
-        project_participation_id: 1,
-        project_id: 2,
-        system_user_id: 10,
-        project_role_id: 2,
-        project_role_name: 'Collaborator'
-      },
-      {
-        project_participation_id: 2,
-        project_id: 2,
-        system_user_id: 20,
-        project_role_id: 2,
-        project_role_name: 'Collaborator'
-      }
-    ];
-
-    const result = delete_endpoint.doAllProjectsHaveAProjectLead(rows);
-
-    expect(result).to.equal(false);
   });
 });
